+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -100,6 +100,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Added mutation methods.
 * Added simple `#define`ed constants.
 * Added `CFString::as_str_unchecked`.
+* Added a few compatibility aliases to help with migrating from the
+  `core-foundation` crate.
 
 ### Changed
 * Made `CFArray`, `CFBag`, `CFBinaryHeap`, `CFDictionary`, `CFSet` and their